Cloud_Billing_Forecast_with_Terraform/requirements.txt
google-cloud-aiplatform 
google-cloud-bigquery
google-cloud-bigquery-storage
google-cloud-storage
pytz
pypi-json
urllib3
